Limit number of retries for gen/such-that
generators/such-that now defaults to only retrying to generate a value
that will pass the predicate 10 times. A two-artiy version is now
available that lets you change the number of retries. You should only
need to do this temporarily, as you figure out a better way to write the
generator. If the predicate doesn't pass after `n` times, an exception
is thrown.