Convert to immutable splittable RNG
Created a protocol for a splittable RNG, in the new
clojure.test.check.random namespace. Also created
an implementation using the same algorithm from
the java.util.SplittableRandom class.

Converted the test runner code to use this, and
adapted a couple tests as necessary.

Wrote tests for the RNG, testing that it's
deterministic and that splitting returns distinct
states. Also a Java-8 only test that checks that
the RNG implementation gives identical results to
equivalent uses of java.util.SplittableRandom.

This change fixes TCHECK-27 (problems stemming from
a non-deterministic shrink-tree), and sets up the
codebase for much easier addition of features such
as:

- Running tests deterministically in parallel from a
  single seed
- Supplying RNG "state" data to reproduce any individual
  test run
- Resuming shrinking
- Supporting alternate shrink-tree-walking algorithms