Add generators for distinct collections.
This wasn't easy, as it had to be done at a low level instead of simply
with pure combinators, because the ideal way to handle duplicate
elements (in a way that interacts with collection sizes and shrinking
well) is at a low level, using the same kind of "timeout" as such-that.

Wrote a generic (private) generator called coll-distinct-by, and then
implemented vector-distinct, list-distinct, set, and sorted-set using
coll-distinct-by. Added similar sizing options to map, by reimplementing
it with coll-distinct-by.

Added tests for distinctness and size both for generated values and
during shrinking.

Also had to complicate a little bit by shuffling the generated
collections to prevent them from biasing towards having smaller values
at the beginning (e.g., without this (gen/vector-distinct gen/nat
{:num-elements 3}) would only generate [0 1 2] with size=0 even though
we'd like all orderings to be equally likely).