Fix issue w/ distinct collection generators
If you gave a distinct collection generator a generator that could only
generate a few things, but didn't give it any particular minimum size,
it could still blow up because it internally picks a particular size
before trying to generate the collection.

In concrete terms, (gen/sample (gen/set gen/boolean)) was highly likely
to fail.

Added a distinction between the target size and the real minimum size,
so we can give up successfully if we can't generate more things but are
still above the real minimum.