Workaround for CLJS-1594: don't hash weird doubles
ClojureScript was having trouble with gen/any and I finally tracked it
down to CLJS-1594, where NaN & the infinities hash to NaN which causes
stack overflows when used in sets and as map keys.

Seemed like the only reasonable course of action was to add filters in
gen/any to make sure we don't put those values in those positions, but
only on the CLJS platform.