TCHECK-105: Support for bootstrapped / self-host ClojureScript
Adds support for bootstrapped / self-hosted ClojureScript.

Additionally, a `script/test-self-host` script is added which facilitates
testing `test.check` when built with self-hosted ClojureScript (this
is done in a self-hosted environment hosted in Node.) This provides
a means for ensuring `test.check` doesn't suffer regressions with
respect to targeting self-hosted ClojureScript.