Improve failure reporting with failing result data and location.
This addresses TCHECK-34 ("reporting is uniformative").

This adds a clojure.test/assert-expr which parses the check results
and emits appropriate messages via clojure.test/do-report. For failed
tests, the arguments to clojure-test/assert-check are returned as the
:actual result, with a placeholder {:result true} as the :expected result.

The file and location provided in ClojureScript environments is not
particularly useful as it reports the file and line location in the
relevant JavaScript file rather than of the source ClojureScript file.

Change-Id: I372c307115d0cc1d62d7460eb253031cd269f750