Dynamically extend cljs.test/assert-expr
The clojure.test.check.clojure-test.assertions.cljs namespace exists
only for the side effect of extending the cljs.test/assert-expr
multimethod.

This has to be done on the clj side of cljs compilation, and so we have
a separate namespace that is only loaded by cljs via a :require-macros
clause in clojure.test.check.clojure-test.assertions. This means we have
a clj namespace that should only be loaded by cljs compilation.

Normally this is not a problem, but it seems that the clojure maven
plugin will eagerly load all source files it can find. I'm not sure why
`cljs.test` is not available in that context, but this seems to fix it.