TCHECK-139: CLJS test warnings re: cljs.test$macros
Avoid spurious warnings that are emitted in the event that
clojure.test.check.clojure-test.assertions.cljs is compiled as a
ClojureScript run time namespace (as opposed to its intended use as a
macros namespace.) This namespace can be inadvertently compiled as a
run time namespace simply because it exists in source paths and is
eligible for compilation (given that it is implemented in a *.cljc
file) and, if this occurs, the ClojureScript analyzer will warn on
undefined Vars.

At a broad level, we need to define the cljs.test/assert-expr
multifunction in three environments: Clojure, JVM ClojureScript, and
self-hosted ClojureScript. The undefined Var warnings are occuring for
the self-hosted ClojureScript definition (when being inadvertently
compiled for JVM ClojureScript.)

This patch takes the approach of using a JavaScript reference to the
multifunction (as opposed to a ClojureScript reference), thus bypassing
the ClojureScript analyzer. The multifunction reference is already
presuming compiler internals, employing the $macros suffix, so this
just goes one step farther and uses a direct JavaScript reference. This
is OK in the face of :advanced renaming, because self-hosted
ClojureScript doesn't support :advanced.

In addition, this patch addresses a secondary issue I noticed when
testing the change: When compiling the definition under self hosted
ClojureScript, it would complain that
clojure.test.check.clojure-test.assertions/check? is undefined. This is
because this Var is defined in the very namespace that is requiring
this macros namespace. The solution is to move the self-hosted
ClojureScript definition out of the macros namespace, and to place it
adjacent to the Clojure definition. Finally, since we only want the
defmethod to be compiled under self-hosted ClojureScript, we guard it
with a check that ensures that js/cljs.test$macros exists.