Provide :test-var value in :complete report only when available.
This fixes TCHECK-140.

When reporting completion results for tests defined via `defspec`,
it's useful to include the name of the test. This is available
in the *testing-vars* value in the test environment when tests
are composed.

When calling the test function directly, *testing-vars* is not
populated with the test-var. It's not necessary to include the
test-var name as its known: it's the function that is being
directly called.