Add size-bounded-bigint to clj-jvm, rewrite ratio in terms of it
The first bigint generator, it explicitly specifies that it's bounded
by a function of the `size` parameter; this is in contrast to earlier
approaches I tried where for any given `size` the distribution would
still be unbounded, but smeared differently.

Describing the distribution in the name leaves room for later adding
other bigint generators that behave differently.

I didn't add it to cljs, since it's unclear what to generate there,
and so ratio uses large-integer instead. It just ends up being doubles
on cljs anyhow so that part doesn't matter much.