Add `-equatable` variants of simple-type and any generators
This is a sad workaround for the eternal problem that collections
containing NaNs, when not `identical?`, cannot be `=`.

The name is more generic as there could conceivably be other objects
added to gen/any & friends that don't qualify for the equable
variants, e.g. (Object.)s.

The word "equatable" could arguably imply the wrong concept of "can be
passed to =", but since everything can be passed to = in clojure (just
like every object implements .equals on the jvm), I'm not worried
about this.