Comment on commit msg 5e2734f6725
Wouldn't it be great if git had a mechanism for annotating old commit
messages or something of that sort?

Anyhow I noticed that commit 5e2734f6725 has an incorrect commit
message due to too much `git commit --amend`ing, and I wanted to
at least say so in the git log.

(So this commit doesn't have any associated changes)

The wrong part is "rewrite ratio in terms of it" -- in reality,
gen/ratio was left alone and gen/big-ratio was added (jvm only),
documented as using `gen/size-bounded-bigint`.